// src/store/popular-collection/types.tsx
import { IPopular } from './reducer'

export enum TYPE {
  POPULAR_COLLECTION_PENDING = 'POPULAR_COLLECTION_PENDING',
  POPULAR_COLLECTION_FAIL = 'POPULAR_COLLECTION_FAIL',
  GET_POPULAR_COLLECTION_ALL = 'GET_POPULAR_COLLECTION_ALL',
  GET_POPULAR_COLLECTION_ID = 'GET_POPULAR_COLLECTION_ID',
}

interface POPULAR_COLLECTION_PENDING {
  type: TYPE.POPULAR_COLLECTION_PENDING
}

interface POPULAR_COLLECTION_FAIL {
  type: TYPE.POPULAR_COLLECTION_FAIL
  payload: string
}

interface GET_POPULAR_COLLECTION_ALL {
  type: TYPE.GET_POPULAR_COLLECTION_ALL
  payload: IPopular[]
  total: number
}

interface GET_POPULAR_COLLECTION_ID {
  type: TYPE.GET_POPULAR_COLLECTION_ID
  payload: number
}

export type Action = POPULAR_COLLECTION_PENDING | POPULAR_COLLECTION_FAIL | GET_POPULAR_COLLECTION_ALL | GET_POPULAR_COLLECTION_ID